+++ NEWS.md
@@ -1,5 +1,6 @@
 # isstatic development
 
+* Rename `is_list_of()` to `is_list_all()` to avoid name conflict with `vctrs::is_list_of()`
 * Add `is_file()`, `is_dir()`, `has_min_length()` and `has_max_length()` functions.
 * Make `str_fileext()` length stable (although functions may now return NA values).